Find all values within a range

This is essentially a shortcut for x >= left & x <= right, but it also retains the size of x and casts both left and right to the type of x before making the comparison.

between(x, left, right, bounds = "[]")

Arguments

x

A vector

left, right

Boundary values. Both left and right are recycled to the size of x and are cast to the type of x.

bounds

One of "[]", "[)", "(]", or "()", which defines whether the boundary is inclusive (with [ or ]) or exclusive (with ( or )).

Value

A logical vector the same size as x.

Examples

between(c(1:10, NA), 4, 6)
#>  [1] FALSE FALSE FALSE  TRUE  TRUE  TRUE FALSE FALSE FALSE FALSE    NA
between(letters, "d", "j")
#>  [1] FALSE FALSE FALSE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE  TRUE FALSE FALSE
#> [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [25] FALSE FALSE

today <- Sys.Date()
between(today, today - 1, today + 1)
#> [1] TRUE