Take a random sample from a vector.

Take a random sample from a vector.

sample(x, size = vec_size(x), replace = FALSE, w = NULL)

Arguments

x

A vector.

size

Number of samples to take, Defaults to the size of x, generating a random permutation.

replace

Should samples be taken with or without replacement?

w

Optionally, a vector of weights. If provided should be a numeric vector the same size as x.

Examples

sample(1:5)
#> [1] 1 4 2 3 5
sample(1:100, 10)
#>  [1] 89 20 14 24 66 84 44 28 68 16
sample(1:10, 100, replace = TRUE)
#>   [1] 10  6  5  7  3  8  9  3  5  6  1  6  4  9  2 10  8  5  7  5  5  7  3  7  9
#>  [26]  1  9 10 10  7 10 10  6  5  2  9  5  1  1 10  3  8  4  1  8  3  6  5  3  3
#>  [51]  3  6  7  7  1  6  6  3 10  1  4  3  7  6  7  3  1  6  4  6  8  1  2 10  9
#>  [76]  5  5 10  2  2  7  3  5 10  3  1  8  9  6  5  8  1  3  4  9  6  3  9  9  2

# samples of data frames are taken row wise, making it easy to
# do bootstrapping
sample(iris, 10, replace = TRUE)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1           5.9         3.2          4.8         1.8 versicolor
#> 2           5.4         3.0          4.5         1.5 versicolor
#> 3           6.2         2.9          4.3         1.3 versicolor
#> 4           5.7         3.0          4.2         1.2 versicolor
#> 5           7.2         3.0          5.8         1.6  virginica
#> 6           4.6         3.4          1.4         0.3     setosa
#> 7           6.9         3.1          4.9         1.5 versicolor
#> 8           6.3         3.4          5.6         2.4  virginica
#> 9           5.1         3.8          1.5         0.3     setosa
#> 10          6.3         2.3          4.4         1.3 versicolor

# Unlike base::sample(), this is safe to use with inputs of length 1
sample(3:5)
#> [1] 4 5 3
sample(4:5)
#> [1] 5 4
sample(5:5)
#> [1] 5
base::sample(5:5)
#> [1] 3 5 2 4 1