Take a random sample from a vector.

Take a random sample from a vector.

sample(x, size = vec_size(x), replace = FALSE, w = NULL)

Arguments

x

A vector.

size

Number of samples to take, Defaults to the size of x, generating a random permutation.

replace

Should samples be taken with or without replacement?

w

Optionally, a vector of weights. If provided should be a numeric vector the same size as x.

Examples

sample(1:5)
#> [1] 3 5 2 1 4
sample(1:100, 10)
#>  [1] 46 25 48 22  9 20 63 82 80 28
sample(1:10, 100, replace = TRUE)
#>   [1]  2  8  8  5  2  5  8 10  9  2  8  9  9  9  1  5  1  3  1  7  6  1  2  9  1
#>  [26]  1  6  6  1  8  9  6  8 10  5  9  5  2  6  7  2  7  4 10  6 10  1  9  8  3
#>  [51]  4  5  3  3  7 10 10  4  1  5  8  9  2  4  3  8  6 10  2  8  1  4  4  9  8
#>  [76]  5  3  8  9 10  8  7 10 10  3  2 10  4  5  2  5  5  3  1  6  5  4  8 10  2

# samples of data frames are taken row wise, making it easy to
# do bootstrapping
sample(iris, 10, replace = TRUE)
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width    Species
#> 1           7.7         3.8          6.7         2.2  virginica
#> 2           7.2         3.2          6.0         1.8  virginica
#> 3           6.6         2.9          4.6         1.3 versicolor
#> 4           6.0         3.0          4.8         1.8  virginica
#> 5           4.3         3.0          1.1         0.1     setosa
#> 6           6.3         2.5          4.9         1.5 versicolor
#> 7           7.9         3.8          6.4         2.0  virginica
#> 8           6.3         2.8          5.1         1.5  virginica
#> 9           5.7         4.4          1.5         0.4     setosa
#> 10          6.1         2.6          5.6         1.4  virginica

# Unlike base::sample(), this is safe to use with inputs of length 1
sample(3:5)
#> [1] 5 3 4
sample(4:5)
#> [1] 5 4
sample(5:5)
#> [1] 5
base::sample(5:5)
#> [1] 2 1 3 4 5